Fix for issue #337
The bug in question where the time value for the cloned page adds a
bunch of numbers to the end (actually the UNIX time) is actually not
SQLite specific.

Due to the way the Page model handles "created_on_time", it appended
the unix time to the end of "created_on". This variable looks to be a
remnant from an earlier one related to the fork from FrogCMS(?).

It only manifested itself in SQLite because SQLite does not enforce
column types strictly, whereas in MySQL the extra numbers are discarded
by the DB, SQLite stores them as a string.

Removing the offending lines fixes the issue and does not change the
functionality of clone().