Fix for issues #305, #324, #355 related to SQLite
The three bugs (and others duplicated such as #356) were all related to
one single problem in the `Framework`, namely that empty values were
not being handled properly for SQLite. Since SQLite does not support
DEFAULT keyword, an empty string must be explicitly passed in order to
properly update the record. The issues mentioned all involved
attempting to insert empty values in the database which was not caught
by the `if` statement on line 590.