Fix issues #305, #324, #355
The three bugs (and others duplicated such as #356) were all related to
one single problem in the Framework, namely that empty values were not
being handled properly for SQLite. Since SQLite does not support the
DEFAULT keyword, an empty string must be explicitly passed in order to
properly update the record. The issues mentioned all involved
attempting to insert empty values in the database which was not caught
by the if statement on line 595.